refatorando codigo

+++ model/Region.py
@@ -1,7 +1,7 @@
 from configuration.config import ormDatabase as orm
  
 class Region(orm.Model):
-    id = orm.Column(orm.BigInteger, primary_key=True)
+    id = orm.Column(orm.Integer, primary_key=True)
     name = orm.Column(orm.String(45))
     
     def __init__(self, name):
