+++ model/State.py
@@ -5,7 +5,7 @@ class State(orm.Model):
     ibge_id = orm.Column(orm.Integer) 
     name = orm.Column(orm.String(100))
     abbreviation = orm.Column(orm.String(2))
-    region_id = orm.Column(orm.BigInteger, orm.ForeignKey('region.id'))
+    region_id = orm.Column(orm.Integer, orm.ForeignKey('region.id'))
     
     def __init__(self, name, abbreviation, region_id):
         self.name = name
